### 2024-06-18 #6900 @danielskatz
👋 @koide3 - thanks for your submission.

Please
1. Work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. After you make changes to your .bib file, then use the command `@editorialbot check references` to check them again
2. Fix the formatting of your references - follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). Use the command `@editorialbot generate pdf` after making changes (or when the references are right) to make a new PDF

`@editorialbot` commands need to be the first entry in a new comment.

In the meantime, I'll look for an editor for your submission.
