+++ docs/paper.bib
@@ -4,7 +4,8 @@
 	booktitle = {{IEEE International Conference on Robotics and Automation (ICRA2011)}},
 	month     = {May},
 	year      = {2011},
-	address   = {Shanghai, China}
+	address   = {Shanghai, China},
+  doi       = {10.1109/ICRA.2011.5980567}
 }
 
 @article{Zhou,
@@ -12,6 +13,7 @@
     title     = {{Open3D}: {A} Modern Library for {3D} Data Processing},
     journal   = {arXiv:1801.09847},
     year      = {2018},
+    doi       = {10.48550/arXiv.1801.09847}
 }
 
 @ARTICLE{Bai,
@@ -21,7 +23,8 @@
   year={2022},
   volume={7},
   number={2},
-  pages={4861-4868}
+  pages={4861-4868},
+  doi={10.1109/LRA.2022.3152830}
 }
 
 @inproceedings{Koide,
@@ -30,7 +33,8 @@
   booktitle  = {IEEE International Conference on Robotics and Automation (ICRA2021)},
   pages      = {11054--11059},
   year       = {2021},
-  month      = {May}
+  month      = {May},
+  doi        = {10.1109/ICRA48506.2021.9560835}
 }
 
 @article{Zhang,
@@ -41,15 +45,17 @@
   number={2},
   pages={119--152},
   year={1994},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/BF01427149}
 }
 
 @inproceedings{Segal,
-  title={Generalized-icp.},
+  title={Generalized-{ICP}},
   author={Segal, Aleksandr and Haehnel, Dirk and Thrun, Sebastian},
   booktitle={Robotics: science and systems},
   volume={2},
   number={4},
   pages={435},
-  year={2009}
+  year={2009},
+  doi={10.15607/rss.2009.v.021}
 }
