### 2024-06-20 0e332e5
parallel batch nearest neighbor search
