### 2024-06-20 7e42a90
parallel batch nearest neighbor search (#68)
