### 2024-06-20 b04752d
improve knn result collection
