### 2024-06-20 fc299d1
optimize expmap
