### 2024-06-21 0d3f5e6
improve knn performance (#70)
* improve knn result collection

* optimize expmap
