### 2024-06-21 0797592
remove unnecessary dependencies for document generation
