+++ .github/workflows/gendoc.yml
@@ -1,0 +1,31 @@
+name: gendoc
+
+on:
+  workflow_dispatch:
+  push:
+    branches:
+      - doci
+    tags:
+      - '*'
+
+jobs:
+  gendoc:
+    name: Generate documentation
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name : Install dependencies
+        run: sudo apt-get install -y git cmake build-essential doxygen python3-dev python3-pip pybind11-dev libeigen3-dev libomp-dev python3-sphinx python3-sphinx-rtd-theme
+
+      - name: Install python dependencies
+        run: python -m pip install mkdocs mkdocs-material
+      
+      -name: Generate documentation
+        run: |
+          cd docs && make all
+      
+      - name: Deploy documentation
+        run: |
+          cd docs && make deploy
