+++ .github/workflows/gendoc.yml
@@ -1,7 +1,6 @@
 name: gendoc
 
 on:
-  workflow_dispatch:
   push:
     branches:
       - doci
