+++ .github/workflows/gendoc.yml
@@ -23,7 +23,7 @@ jobs:
           git config user.email 41898282+github-actions[bot]@users.noreply.github.com
 
       - name : Install dependencies
-        run: sudo apt-get install -y git cmake build-essential doxygen python3-dev python3-pip pybind11-dev libeigen3-dev libomp-dev python3-sphinx python3-sphinx-rtd-theme
+        run: sudo apt-get install -y git cmake build-essential doxygen graphviz python3-dev python3-pip pybind11-dev libeigen3-dev libomp-dev python3-numpy python3-sphinx python3-sphinx-rtd-theme
 
       - name: Install python dependencies
         run: python -m pip install mkdocs mkdocs-material
