### 2024-06-21 f48faf0
Auto generate and deploy documentation (#71)
* show constructor docs

* constructor docs

* remove unnecessary dependencies for document generation

* trigger on tags
