+++ src/python/align.cpp
@@ -99,11 +99,11 @@ void define_align(py::module& m) {
 
         Parameters
         ----------
-        target_points : NDArray[np.float64]
+        target_points : numpy.ndarray[np.float64]
             Nx3 or Nx4 matrix representing the target point cloud.
-        source_points : NDArray[np.float64]
+        source_points : numpy.ndarray[np.float64]
             Nx3 or Nx4 matrix representing the source point cloud.
-        init_T_target_source : np.ndarray[np.float64]
+        init_T_target_source : numpy.ndarray[np.float64]
             4x4 matrix representing the initial transformation from target to source.
         registration_type : str = 'GICP'
             Type of registration algorithm to use ('ICP', 'PLANE_ICP', 'GICP', 'VGICP').
@@ -117,9 +117,10 @@ void define_align(py::module& m) {
             Number of threads to use for parallel processing.
         max_iterations : int = 20
             Maximum number of iterations for the optimization algorithm.
+
         Returns
         -------
-        RegistrationResult
+        result : RegistrationResult
             Object containing the final transformation matrix and convergence status.
         )pbdoc");
 
@@ -173,7 +174,7 @@ void define_align(py::module& m) {
         Pointer to the source point cloud.
     target_tree : KdTree<PointCloud>::ConstPtr, optional
         Pointer to the KD-tree of the target for nearest neighbor search. If nullptr, a new tree is built.
-    init_T_target_source : NDArray[np.float64]
+    init_T_target_source : numpy.ndarray[np.float64]
         4x4 matrix representing the initial transformation from target to source.
     registration_type : str = 'GICP'
         Type of registration algorithm to use ('ICP', 'PLANE_ICP', 'GICP').
@@ -183,9 +184,10 @@ void define_align(py::module& m) {
         Number of threads to use for computation.
     max_iterations : int = 20
         Maximum number of iterations for the optimization algorithm.
+
     Returns
     -------
-    RegistrationResult
+    result : RegistrationResult
         Object containing the final transformation matrix and convergence status.
     )pbdoc");
 
@@ -221,7 +223,7 @@ void define_align(py::module& m) {
         Voxel map constructed from the target point cloud.
     source : PointCloud
         Source point cloud to align to the target.
-    init_T_target_source : NDArray[np.float64]
+    init_T_target_source : numpy.ndarray[np.float64]
         4x4 matrix representing the initial transformation from target to source.
     max_correspondence_distance : float = 1.0
         Maximum distance for corresponding point pairs.
@@ -232,7 +234,7 @@ void define_align(py::module& m) {
 
     Returns
     -------
-    RegistrationResult
+    result : RegistrationResult
         Object containing the final transformation matrix and convergence status.
     )pbdoc");
 }
