+++ src/python/kdtree.cpp
@@ -18,18 +18,50 @@ void define_kdtree(py::module& m) {
   // KdTree
   py::class_<KdTree<PointCloud>, std::shared_ptr<KdTree<PointCloud>>>(m, "KdTree")  //
     .def(
+      py::init([](const Eigen::MatrixXd& points_numpy, int num_threads) {
+        if (points_numpy.cols() != 3 && points_numpy.cols() != 4) {
+          throw std::invalid_argument("points must have shape (n, 3) or (n, 4)");
+        }
+
+        auto points = std::make_shared<PointCloud>();
+        points->resize(points_numpy.rows());
+
+        for (int i = 0; i < points_numpy.rows(); i++) {
+          points->point(i) << points_numpy(i, 0), points_numpy(i, 1), points_numpy(i, 2), 1.0;
+        }
+
+        return std::make_shared<KdTree<PointCloud>>(points, KdTreeBuilderOMP(num_threads));
+      }),
+      py::arg("points"),
+      py::arg("num_threads") = 1,
+      R"""(
+        KdTree(points: numpy.ndarray, num_threads: int = 1)
+
+        Construct a KdTree from numpy.
+
+        Parameters
+        ----------
+        points : numpy.ndarray, shape (n, 3) or (n, 4)
+            The input point cloud.
+        num_threads : int, optional
+            The number of threads to use for KdTree construction. Default is 1.
+       )""")
+
+    .def(
       py::init([](const PointCloud::ConstPtr& points, int num_threads) { return std::make_shared<KdTree<PointCloud>>(points, KdTreeBuilderOMP(num_threads)); }),
       py::arg("points"),
       py::arg("num_threads") = 1,
       R"""(
-         Construct a KdTree from a point cloud.
+        KdTree(points: PointCloud, num_threads: int = 1)
 
-         Parameters
-         ----------
-         points : PointCloud
-             The input point cloud.
-         num_threads : int, optional
-             The number of threads to use for KdTree construction. Default is 1.
+        Construct a KdTree from a point cloud.
+
+        Parameters
+        ----------
+        points : PointCloud
+            The input point cloud.
+        num_threads : int, optional
+            The number of threads to use for KdTree construction. Default is 1.
        )""")
 
     .def(
@@ -42,21 +74,21 @@ void define_kdtree(py::module& m) {
       },
       py::arg("pt"),
       R"""(
-         Find the nearest neighbor to a given point.
+        Find the nearest neighbor to a given point.
 
-         Parameters
-         ----------
-         pt : NDArray, shape (3,)
-             The input point.
+        Parameters
+        ----------
+        pt : numpy.ndarray, shape (3,)
+            The input point.
 
-         Returns
-         -------
-         found : int
-             Whether a neighbor was found (1 if found, 0 if not).
-         k_index : int
-             The index of the nearest neighbor in the point cloud.
-         k_sq_dist : float
-             The squared distance to the nearest neighbor.
+        Returns
+        -------
+        found : int
+            Whether a neighbor was found (1 if found, 0 if not).
+        k_index : int
+            The index of the nearest neighbor in the point cloud.
+        k_sq_dist : float
+            The squared distance to the nearest neighbor.
        )""")
 
     .def(
@@ -70,21 +102,21 @@ void define_kdtree(py::module& m) {
       py::arg("pt"),
       py::arg("k"),
       R"""(
-       Find the k nearest neighbors to a given point.
+        Find the k nearest neighbors to a given point.
 
-       Parameters
-       ----------
-       pt : NDArray, shape (3,)
-           The input point.
-       k : int
-           The number of nearest neighbors to search for.
+        Parameters
+        ----------
+        pt : numpy.ndarray, shape (3,)
+            The input point.
+        k : int
+            The number of nearest neighbors to search for.
 
-       Returns
[…]
