### 2024-06-21 a39ce13
octree (just for experiments)
