+++ include/small_gicp/ann/octree.hpp
@@ -1,0 +1,215 @@
+#pragma once
+
+#include <random>
+#include <numeric>
+#include <small_gicp/points/traits.hpp>
+#include <small_gicp/ann/knn_result.hpp>
+
+namespace small_gicp {
+
+struct OctreeNode {
+  using NodeIndexType = std::uint32_t;
+  static constexpr NodeIndexType INVALID_NODE = std::numeric_limits<NodeIndexType>::max();
+  static constexpr NodeIndexType LEAF_NODE = INVALID_NODE - 1;
+
+  union {
+    struct Leaf {
+      NodeIndexType is_leaf;
+      NodeIndexType first;
+      NodeIndexType last;
+    } lr;
+    struct NonLeaf {
+      std::array<NodeIndexType, 8> children;
+      std::array<double, 4> separator;
+    } sub;
+  } node_type;
+};
+
+template <typename PointCloud>
+struct Octree {
+public:
+  Octree(const PointCloud& points) : points(points) {
+    Eigen::Vector4d min_pt = traits::point(points, 0);
+    Eigen::Vector4d max_pt = traits::point(points, 0);
+    for (size_t i = 1; i < points.size(); ++i) {
+      min_pt = min_pt.cwiseMin(traits::point(points, i));
+      max_pt = max_pt.cwiseMax(traits::point(points, i));
+    }
+
+    origin = min_pt;
+    extent = max_pt - min_pt;
+
+    indices.resize(points.size());
+    std::iota(indices.begin(), indices.end(), 0);
+
+    size_t node_count = 0;
+    nodes.resize(points.size() * 2);
+    root = create_node(origin, extent, node_count, indices.data(), indices.data(), indices.data() + points.size());
+
+    nodes.resize(node_count);
+  }
+
+  size_t knn_search(const Eigen::Vector4d& query, int k, size_t* k_indices, double* k_sq_dists, const KnnSetting& setting = KnnSetting()) const {
+    KnnResult<-1> result(k_indices, k_sq_dists, k);
+    knn_search(query, root, result, setting);
+    return result.num_found();
+  }
+
+public:
+  int calc_child_index(const Eigen::Vector4d& separator, const Eigen::Vector4d& query) const {
+    const Eigen::Matrix<bool, 4, 1> comp = query.array() > separator.array();
+    return comp[0] + comp[1] * 2 + comp[2] * 4;
+  }
+
+  Eigen::Vector4d find_best_separator(const Eigen::Vector4d& origin, const Eigen::Vector4d& extent, size_t* first, size_t* last) {
+    const auto evaluate = [&](const Eigen::Vector4d& separator, std::array<size_t, 8>& counts) {
+      std::fill(counts.begin(), counts.end(), 0);
+      for (size_t* it = first; it != last; ++it) {
+        const int index = calc_child_index(separator, traits::point(points, *it));
+        counts[index]++;
+      }
+
+      double entropy = 0.0;
+      for (size_t count : counts) {
+        if (count != 0) {
+          const double p = static_cast<double>(count) / std::distance(first, last);
+          entropy -= p * std::log(p);
+        }
+      }
+
+      return entropy;
+    };
+
+    Eigen::Vector4d baseline_separator = origin + extent * 0.5;
+    std::array<size_t, 8> baseline_counts;
+    double baseline_entropy = evaluate(baseline_separator, baseline_counts);
+
+    Eigen::Vector4d best_separator = baseline_separator;
+    std::array<size_t, 8> best_counts = baseline_counts;
+    double best_entropy = baseline_entropy;
+
+    std::mt19937 mt(*first + (origin.array() * Eigen::Array4d(10231, 12321, 8412541, 0.0)).sum());
+    std::uniform_real_distribution<> udist(0.0, 1.0);
+
[…]
