### 2024-06-27 e95cb19
Fix BUILD_WITH_OPENMP being set to an incorrect value by default (#76)
