### 2024-06-27 7a838f1
expose verbose option to python
