### 2024-06-27 e1f77dd
tweak convergence check
