+++ .github/workflows/license.yml
@@ -10,14 +10,14 @@ on:
 
 jobs:
   license_check:
-    name: License indication check
+    name: License check
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
       - name: Check header files
-        run: "find include/small_gicp -type f | xargs -I filename bash -c 'if ! grep -q SPDX-License-Identifier: MIT filename; then echo filename : lisence not found; exit 1; fi'"
+        run: "find include/small_gicp -type f | xargs -I filename bash -c 'if ! grep -q SPDX-License-Identifier filename; then echo filename : lisence not found; exit 1; fi'"
 
       - name: Check source files
-        run: "find src/small_gicp -type f | xargs -I filename bash -c 'if ! grep -q SPDX-License-Identifier: MIT filename; then echo filename : lisence not found; exit 1; fi'"
+        run: "find src/small_gicp -type f | xargs -I filename bash -c 'if ! grep -q SPDX-License-Identifier filename; then echo filename : lisence not found; exit 1; fi'"
