+++ include/small_gicp/factors/robust_kernel.hpp
@@ -1,3 +1,5 @@
+// SPDX-FileCopyrightText: Copyright 2024 Kenji Koide
+// SPDX-License-Identifier: MIT
 #pragma once
 
 #include <Eigen/Core>
