+++ .github/workflows/license.yml
@@ -2,7 +2,7 @@ name: license
 
 on:
   push:
-    branches: [ master, license-check ]
+    branches: [ master ]
     paths-ignore: '**.md'
   pull_request:
     branches: [ master ]
