### 2024-06-27 45b0b29
expose verbose option to python (#77)
* expose verbose option to python

* tweak convergence check
