### 2024-06-27 765da6f
CI for license check (#78)
* license check CI

* fix ci

* check both headers and source

* add SPDX identifiers

* trigger on only master
