+++ doc_cpp/search/variables_1.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['b_564',['b',['../structsmall__gicp_1_1LinearizeSum.html#a7c6ac278a4ce034f5780706d5064cd96',1,'small_gicp::LinearizeSum::b()'],['../structsmall__gicp_1_1RegistrationResult.html#a8f361198c487333477a8dc0e0efd7991',1,'small_gicp::RegistrationResult::b()']]]
+];
