+++ doc_cpp/search/all_1.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['b_4',['b',['../structsmall__gicp_1_1LinearizeSum.html#a7c6ac278a4ce034f5780706d5064cd96',1,'small_gicp::LinearizeSum::b()'],['../structsmall__gicp_1_1RegistrationResult.html#a8f361198c487333477a8dc0e0efd7991',1,'small_gicp::RegistrationResult::b()']]],
+  ['buffer_5fsize_5',['buffer_size',['../structsmall__gicp_1_1KnnResult.html#a2789fb872207d6955c18f901e2957c14',1,'small_gicp::KnnResult']]],
+  ['build_5ftree_6',['build_tree',['../structsmall__gicp_1_1KdTreeBuilder.html#aebbc718d567041dbb1513d09d0a428d6',1,'small_gicp::KdTreeBuilder::build_tree()'],['../structsmall__gicp_1_1KdTreeBuilderOMP.html#ab0c657027d04b3f776bbdabdb250d0ab',1,'small_gicp::KdTreeBuilderOMP::build_tree()'],['../structsmall__gicp_1_1KdTreeBuilderTBB.html#a41605c07bc9621f5f5054547d9991094',1,'small_gicp::KdTreeBuilderTBB::build_tree()']]]
+];
