+++ doc_cpp/search/classes_d.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['serialreduction_373',['SerialReduction',['../structsmall__gicp_1_1SerialReduction.html',1,'small_gicp']]],
+  ['setting_374',['Setting',['../structsmall__gicp_1_1Cauchy_1_1Setting.html',1,'small_gicp::Cauchy::Setting'],['../structsmall__gicp_1_1FlatContainer_1_1Setting.html',1,'small_gicp::FlatContainer&lt; HasNormals, HasCovs &gt;::Setting'],['../structsmall__gicp_1_1GaussianVoxel_1_1Setting.html',1,'small_gicp::GaussianVoxel::Setting'],['../structsmall__gicp_1_1GICPFactor_1_1Setting.html',1,'small_gicp::GICPFactor::Setting'],['../structsmall__gicp_1_1Huber_1_1Setting.html',1,'small_gicp::Huber::Setting'],['../structsmall__gicp_1_1ICPFactor_1_1Setting.html',1,'small_gicp::ICPFactor::Setting'],['../structsmall__gicp_1_1PointToPlaneICPFactor_1_1Setting.html',1,'small_gicp::PointToPlaneICPFactor::Setting'],['../structsmall__gicp_1_1RobustFactor_1_1Setting.html',1,'small_gicp::RobustFactor&lt; Kernel, Factor &gt;::Setting']]]
+];

[…]
