### 2024-07-01 b1d84b0
improve knn performance of voxelmaps
