### 2024-07-01 37fb6a8
add voxel search patterns
