### 2024-07-01 b2bed0c
add (gicp|vgicp)_model_omp
