### 2024-07-01 76b2f00
improve knn performance of voxelmaps (#79)
* improve knn performance of voxelmaps

* add voxel search patterns

* add (gicp|vgicp)_model_omp
