### 2024-07-01 #6948 @editorialbot
[REVIEW]: small_gicp: Efficient and parallel algorithms for point cloud registration
**Submitting author:** @koide3 (<a href="http://orcid.org/0000-0001-5361-1428">Kenji Koide</a>)
**Repository:** https://github.com/koide3/small_gicp
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
