### 2024-07-02 #6948 @koide3
@versatran01  @abougouffa
Thank you so much for reviewing small_gicp!
I'm so glad that my package is going to be reviewed by professionals in the community.
I'm looking forward to hearing any feedback from you.
