### 2024-07-21 #6948 @abougouffa
Hi @diehlpk

I've read the paper and taken a look at the code but didn't have the time for now to check the code in deep and try it locally. I will try to finish the review next week!

Thanks for your patience!
