### 2024-07-28 #6948 @abougouffa
I would like to thank you for your patience, and thank @koide3 for the excellent implementation.

I have finished the review of the paper and the code, the code is well written and documented. The installation and execution went as described in the documentation. The results I got by running the benchmarks on the same dataset used by the author gave results similar to these presented in the repository.

The paper might need few modifications:

> Iterative point cloud registration, also known as fine registration or local registration, ***is particularly crucial***.

Besides introducing the "iterative point cloud registration" term and its aliases, this phrase doesn't add any significant information. I would like to have a phrase like:

> Iterative point cloud registration, also known as fine registration or local registration, is a _<<some relevant definition of the ***iterative*** aspect of "iterative point cloud registration">>_.

Also, I think you can add some citations of papers using the iterative point cloud registration to solve these problems:

> They are widely used in applications like autonomous vehicle localization [CITATION], place recognition [CITATION], and object classification [CITATION].

In the statement of needs:

> There are several point cloud processing libraries, ~and~ **[other conjunction might be better!]** PCL (Rusu & Cousins, 2011) and Open3D (Zhou et al., 2018) are ***the notable ones among them***.

The ***"notable"*** qualification here seems a bit arbitrary! Indeed, there are other known open source libraries like [`libpointmatcher`](https://github.com/norlab-ulaval/libpointmatcher), [`CGAL`](https://github.com/cgal/cgal), or the [`CCCoreLib`](https://github.com/CloudCompare/CCCoreLib) used in [CloudCompare](https://www.cloudcompare.org/). So, I think the phrase need to be changed to something more neutral.
