### 2024-08-04 #6948 @versatran01
I have reviewed the paper and the code and checked all checklist items.

* **Functionality[Installation]/Documentation[Installation Instructions]**: I was able to follow the installation guidelines and successfully compiled the code in a conda environment where all dependencies are installed via conda. I was also able to build Iridescence but after installation, small_gicp couldn't find the library. But this was not an issue with the code under review.

* **Functionality[Functionality]/Documentation[Example Usage]**: I have no trouble running the examples 01, 02, 03.

* **Functionality[Performance]**: I could not run all the benchmarks because it requires the kitti dataset and is a fairly big dataset to download. However, the author did provide detailed benchmark results in the repo. By looking at the code, both the downsampling and ICP algorithms are accelerated by multithreading, and the alleged speed-up does make sense. I would suggest the author either provide a smaller subset of the kitti dataset (if allowed) or another smaller dataset to download for the ease of reproducibility.

* **Documentation[Functionality Documentation]**: This is one of the parts where the repo could use some improvements. While most of the functions have a doxygen-style comment, some of the details are missing. For example, Here is the documentation of the `voxelgrid_sampling` function. It has a brief comment that says Voxelgrid downsampling and a note says that discretized voxel coords must be in 21-bit range. However, it is missing some important details (even though many could be inferred by the function signature or the code). For example, what happens to the points outside the 21-bit range? Is this downsampling stable? How many points are kept per voxel? These should be documented. The same concern applies to many other functions in the repo. To summarize, I think the documentation part could use some improvements.

* **Documentation[Automated tests]**: I was able to run all the tests with no issues (just by running `make test`). And the repo contains a healthy amount of tests. However, I'm not able to run code coverage locally via `make coverage`. There's some build failure.

* **Paper[Reference]**: The code and paper already covered many practical use cases of GICP with good references. For example, data association with VoxelHash and KD-tree and various ICP implementations. However, I think in future work, the author could potentially add projective-based ICP, as described in the NICP paper by Serafin and Grisetti.

Overall I think this is good work and a valuable contribution to the robotics community.
