### 2024-08-05 #6948 @koide3
@versatran01
Thank you for reviewer our paper and code!

> I was also able to build Iridescence but after installation, small_gicp couldn't find the library.

It seems some changes in Iridescence introduced cmake errors. This should be fixed quickly.

> I would suggest the author either provide a smaller subset of the kitti dataset (if allowed) or another smaller dataset to download for the ease of reproducibility.

I agree. I think it is possible to distribute a part of KITTI because it employs CC BY-NC-SA license that allows re-distribution. I'll upload a small part of KITTI so that the user can easily reproduce the evaluation results.

> To summarize, I think the documentation part could use some improvements.

I also agree with this point. I'll enhance documentation to tell the detailed behavior of the implemented algorithms to the user.

> However, I think in future work, the author could potentially add projective-based ICP, as described in the NICP paper by Serafin and Grisetti.

Thank you for your suggestion. I'll revise the manuscript with these points along with suggestions given by @abougouffa .

@abougouffa @versatran01
I'll update both the documentation and manuscript and let you know in this week. Thank you again for your suggestions that are so helpful and useful to improve the quality of our package.
