### 2024-08-05 e3f5ba2
detailed documentation (cpp)
