### 2024-08-05 f6a48e5
doc for invalid normal estimation results
