### 2024-08-05 57e841d
docs of detailed algorithm behavior for py
