### 2024-08-06 0a2617d
Documentation of detailed behaviors (#82)
* detailed documentation (cpp)

* doc for invalid normal estimation results

* docs of detailed algorithm behavior for py
