### 2024-08-06 #6948 @koide3
Thank you again for taking the time to review small_gicp! I revised the code, doc, and manuscript as follows. I hope the quality of the package and paper are greatly improved thanks to your suggestions. Please let me know if there are areas that need more improvements.

@abougouffa

> Besides introducing the "iterative point cloud registration" term and its aliases, this phrase doesn't add any significant information. I would like to have a phrase like:

I revised the sentence as follows to clarify what "iterative point cloud registration" means.

`Iterative point cloud registration, also known as fine registration or local registration, iteratively refines the transformation between point clouds starting from an initial guess. Each iteration involves a proximity-based point correspondence search and the minimization of the distance between corresponding points, continuing until convergence.`

---
> Also, I think you can add some citations of papers using the iterative point cloud registration to solve these problems:

I added relevant references as follows.

`They are widely used in applications like autonomous vehicle localization [@Kim], place recognition [@Wang], and object classification [@Izadinia].`

---
> The "notable" qualification here seems a bit arbitrary! Indeed, there are other known open source libraries like [libpointmatcher](https://github.com/norlab-ulaval/libpointmatcher), [CGAL](https://github.com/cgal/cgal), or the [CCCoreLib](https://github.com/CloudCompare/CCCoreLib) used in [CloudCompare](https://www.cloudcompare.org/). So, I think the phrase need to be changed to something more neutral.

I agree, the use of "notable" was a bit subjective. I revised the sentence as follows to put emphasis on processing speed and highlight libraries that are frequently used in real-time applications.

`There are several point cloud processing libraries, and PCL [@Rusu], Open3D [@Zhou], libpointmatcher [@Pomerleau] are commonly used in real-time applications owing to their performant implementations.`

==================

@versatran01

> I was also able to build Iridescence but after installation, small_gicp couldn't find the library.

I fixed the build error with https://github.com/koide3/small_gicp/pull/81

---
> I would suggest the author either provide a smaller subset of the kitti dataset (if allowed) or another smaller dataset to download for the ease of reproducibility.

I uploaded a small subset of KITTI00 sequence and added a link to the benchmark page: https://github.com/koide3/small_gicp/blob/master/BENCHMARK.md#dataset

---
> However, it is missing some important details (even though many could be inferred by the function signature or the code).

I updated both C++ and python documentations to tell detailed behaviors of the implemented algorithms. If I'm still missing anything to be noted, please let me know.

Updates: https://github.com/koide3/small_gicp/pull/82
Updated doc: https://koide3.github.io/small_gicp/doc_py/index.html#small_gicp.voxelgrid_sampling

---
> However, I'm not able to run code coverage locally via make coverage. There's some build failure.

I guess it stems from some dependency problems. While I think it is a minor issue (we can at least check the coverage through CI), if you recommend addressing this, I'll try to make `coverage` runnable on different ubuntu versions.

---
> However, I think in future work, the author could potentially add projective-based ICP, as described in the NICP paper by Serafin and Grisetti.

Thank you for your suggestion! Additional neighbor search methods must be beneficial. I think we can easily integrate new neighbor search methods into small_gicp thanks to its design, and I'm planning to add a few implementations including the projective search. I added a future work section in the paper as follows.

`The efficiency of nearest neighbor search significantly impacts the overall performance of point cloud registration.
While small_gicp currently offers efficient and parallel implementations of KdTree and voxelmap, which are general and useful in many situations, there are other nearest neighbor search methods that can be more efficient under mild assumptions about the point cloud measurement model (e.g., projective search [@Serafin]). We plan to implement these alternative neighbor search algorithms to further enhance the speed of the point cloud registration process. The design of small_gicp, where nearest neighbor search and pose optimization are decoupled, facilitates the easy integration of these new search algorithms.`
