+++ docs/paper.bib
@@ -74,7 +74,8 @@
   author={Izadinia, Hamid and Seitz, Steven M},
   booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
   pages={930--939},
-  year={2020}
+  year={2020},
+  doi={10.1109/cvpr42600.2020.00101}
 }
 
 @ARTICLE{Kim,
