+++ docs/paper.bib
@@ -72,7 +72,7 @@
 @inproceedings{Izadinia,
   title={Scene recomposition by learning-based icp},
   author={Izadinia, Hamid and Seitz, Steven M},
-  booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
+  booktitle={IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR2020)},
   pages={930--939},
   year={2020},
   doi={10.1109/cvpr42600.2020.00101}
