### 2024-08-08 #6948 @diehlpk
@koide3 here are some [examples](https://isocpp.org/wiki/faq/misc-environmental-issues#latex-macros) to make C++ looks nicer in the PDF. You might want to consider that.
