### 2024-08-09 8cbec29
Add paper (#84)
* add paper.(md|bib)

* Update paper.yml

* fix references

* revise paper

* add missing DOI

* consistent citation style

* use CC and url commands
