+++ doc_cpp/search/all_10.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['quick_5fsort_5fomp_212',['quick_sort_omp',['../namespacesmall__gicp.html#a96f8d8685abd1a79630e0ca7b34bd50f',1,'small_gicp']]],
+  ['quick_5fsort_5fomp_5fimpl_213',['quick_sort_omp_impl',['../namespacesmall__gicp.html#a5d3e11647aed6d5f464a3daf45586a1b',1,'small_gicp']]]
+];
