+++ doc_cpp/search/classes_6.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['has_5fnearest_5fneighbor_5fsearch_345',['has_nearest_neighbor_search',['../structsmall__gicp_1_1traits_1_1has__nearest__neighbor__search.html',1,'small_gicp::traits']]],
+  ['huber_346',['Huber',['../structsmall__gicp_1_1Huber.html',1,'small_gicp']]]
+];
