+++ doc_cpp/search/functions_9.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['join_493',['join',['../structsmall__gicp_1_1LinearizeSum.html#a008b56793a392fb1f9b930be59be6bbf',1,'small_gicp::LinearizeSum::join()'],['../structsmall__gicp_1_1ErrorSum.html#a76119fdd779f875513d1e6237706f791',1,'small_gicp::ErrorSum::join()']]]
+];
