### 2024-08-09 #6948 @koide3
I confirmed the authors and affiliations are correct.

I made a release "v1.0.0": https://github.com/koide3/small_gicp/releases/tag/v1.0.0

I archived the release on Zenodo: [10.5281/zenodo.13283012](https://doi.org/10.5281/zenodo.13283012)

I edited the title and the author list of the archive to match them and those in the paper.

The license in the archive is changed to MIT that is the same as software license.
