### 2024-08-09 #6948 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
