### 2024-08-09 #6948 @danielskatz
👋 @koide3 - I just see a couple of small things in the cases of characters in the references, which I've suggested in https://github.com/koide3/small_gicp/pull/85. Please merge this, or let me know what you disagree with, then we can continue.
