### 2024-08-09 b2e9ebd
Update paper.bib (#85)
