+++ docs/paper.bib
@@ -1,6 +1,6 @@
 @InProceedings{Rusu,
 	author    = {Radu Bogdan Rusu and Steve Cousins},
-	title     = {{3D is here: Point Cloud Library (PCL)}},
+	title     = {{3D} is here: Point Cloud Library {(PCL)}},
 	booktitle = {{IEEE International Conference on Robotics and Automation (ICRA2011)}},
 	month     = {May},
 	year      = {2011},
@@ -70,7 +70,7 @@
 }
 
 @inproceedings{Izadinia,
-  title={Scene recomposition by learning-based icp},
+  title={Scene recomposition by learning-based {ICP}},
   author={Izadinia, Hamid and Seitz, Steven M},
   booktitle={IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR2020)},
   pages={930--939},
