+++ docs/paper.bib
@@ -91,7 +91,7 @@
 
 @article{Pomerleau,
 	author = {Pomerleau, Fran{\c c}ois and Colas, Francis and Siegwart, Roland and Magnenat, St{\'e}phane},
-	title = {{Comparing ICP Variants on Real-World Data Sets}},
+	title = {Comparing {ICP} Variants on Real-World Data Sets},
 	journal = {Autonomous Robots},
 	year = {2013},
 	volume = {34},
