### 2024-08-09 #6948 @danielskatz
@koide3 - sorry, I missed one fix that I now see: https://github.com/koide3/small_gicp/pull/86
