### 2024-08-10 #6948 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
