### 2024-08-10 #6948 @danielskatz
Congratulations to @koide3 (Kenji Koide) on your publication!!

JOSS depends on volunteers; we couldn't do this without you!
